+++ src/TwigComponent/src/ComponentAttributes.php
@@ -21,7 +21,7 @@ use Symfony\WebpackEncoreBundle\Dto\AbstractStimulusDto;
  */
 final class ComponentAttributes implements \Stringable, \IteratorAggregate, \Countable
 {
-    private const NESTED_REGEX = '#^([\w-]+):(.+)$#';
+    private const NESTED_REGEX = '#^([\w-]+):(.+)(?<!!)$#';
 
     /** @var array<string,true> */
     private array $rendered = [];
@@ -62,6 +62,10 @@ final class ComponentAttributes implements \Stringable, \IteratorAggregate, \Cou
 
                 if (true === $value && str_starts_with($key, 'aria-')) {
                     $value = 'true';
+                }
+
+                if (str_ends_with($key, '!')) {
+                    $key = substr($key, 0, -1);
                 }
 
                 return match ($value) {